Refactor Bar Chart Markers.
